remove lib/info.dart
